Bench 6550747

+++ src/ttable.h
@@ -48,7 +48,7 @@ void InitHashTable(S_HashTable* table, uint64_t MB);
 
 [[nodiscard]] bool ProbeHashEntry(const ZobristKey posKey, S_HashEntry* tte);
 
-void StoreHashEntry(const ZobristKey key, const int16_t move, int score, int16_t eval, const int bound, const int depth, const bool pv, const bool wasPV);
+void StoreHashEntry(const ZobristKey key, const int16_t move, int score, int eval, const int bound, const int depth, const bool pv, const bool wasPV);
 
 [[nodiscard]] uint64_t Index(const ZobristKey posKey);
